[SyncFS] Replace TrackerSet with TrackerIDSet
Introduce TrackerIDSet and replace TrackerSet with it.
Unlike TrackerSet, TrackerIDSet doesn't have raw pointer to FileTrackers.
The replacement hides FileTrackers from the user of the class.

Review URL: https://codereview.chromium.org/171773010

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@252555 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/sync_file_system/drive_backend/uninstall_app_task.cc
@@ -14,7 +14,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include "chrome/browser/sync_file_system/drive_backend/metadata_database.h"
 #include "chrome/browser/sync_file_system/drive_backend/metadata_database.pb.h"
 #include "chrome/browser/sync_file_system/drive_backend/sync_engine_context.h"
-#include "chrome/browser/sync_file_system/drive_backend/tracker_set.h"
+#include "chrome/browser/sync_file_system/drive_backend/tracker_id_set.h"
 #include "chrome/browser/sync_file_system/syncable_file_system_util.h"
 #include "google_apis/drive/drive_api_parser.h"
 #include "google_apis/drive/gdata_wapi_parser.h"
@@ -48,7 +48,7 @@ void UninstallAppTask::Run(const SyncStatusCallback& callback) {
   DCHECK_EQ(RemoteFileSyncService::UNINSTALL_AND_PURGE_REMOTE, uninstall_flag_);
 
   int64 sync_root_tracker_id = metadata_database()->GetSyncRootTrackerID();
-  TrackerSet trackers;
+  TrackerIDSet trackers;
   if (!metadata_database()->FindTrackersByParentAndTitle(
           sync_root_tracker_id, app_id_, &trackers) ||
       !trackers.has_active()) {
@@ -56,12 +56,18 @@ void UninstallAppTask::Run(const SyncStatusCallback& callback) {
     return;
   }
 
-  FileTracker* app_root_tracker = trackers.active_tracker();
-  app_root_tracker_id_ = app_root_tracker->tracker_id();
-  DCHECK(app_root_tracker->has_synced_details());
+  FileTracker app_root_tracker;
+  if (!metadata_database()->FindTrackerByTrackerID(
+          trackers.active_tracker(), &app_root_tracker)) {
+    NOTREACHED();
+    callback.Run(SYNC_STATUS_FAILED);
+    return;
+  }
+  app_root_tracker_id_ = app_root_tracker.tracker_id();
+  DCHECK(app_root_tracker.has_synced_details());
 
   drive_service()->DeleteResource(
-      app_root_tracker->file_id(),
+      app_root_tracker.file_id(),
       std::string(),  // etag
       base::Bind(&UninstallAppTask::DidDeleteAppRoot,
                  weak_ptr_factory_.GetWeakPtr(),
